third_party/blink/web_tests/external/wpt/loading/lazyload/image-loading-lazy-in-viewport-dynamic.tentative-expected.txt

main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
main frame - didReceiveTitle: In viewport images with loading='lazy' and changed to loading='eager' do not block the window load event
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
This is a testharness.js-based test.
FAIL Test that in viewport images with loading='lazy' and changed to loading='eager' do not block the window load event. assert_true: The window.load() event should fire before in_viewport image loads. expected true got false
Harness: the test ran to completion.
